+++ NEWS.md
@@ -1,3 +1,5 @@
+# chirps (development version)
+
 chirps 0.1.2 (2020-07-12)
 =========================