+++ NEWS.md
@@ -35,7 +35,7 @@ chirps 0.0.6 (2020-01-29)
 
 * Comments/suggestions given by Claudia Vitolo in rOpenSci pkg review are added
 * `dataframe_to_geojson()`, `sf_to_geojson()` are added as exported functions avoiding `chirps:::`
-* documentation for `tapajos` is given avoiding chirps:::
+* documentation for `tapajos` is given avoiding `chirps:::`
 
 
 chirps 0.0.5 (2020-01-09)