+++ NEWS.md
@@ -1,14 +1,14 @@
 chirps 0.1.3 (2021-07-01)
 =========================
 
-# ENHANCEMENTS 
+### ENHANCEMENTS 
 
 * Add new function `get_chirts()` to fetch temperature data from CHC server (https://data.chc.ucsb.edu/products/CHIRTSdaily/v1.0/global_cogs_p05/)
 * Implement data fetching from CHC server in `get_chirps()` which offers a better alternative for requests with multiple data points using GoC files from CHC server (https://data.chc.ucsb.edu/products/CHIRPS-2.0/global_daily/cogs/) and the `terra` package  
 * New S3 methods in `get_chirps()` for objects of class 'SpatVector' and 'SpatRaster' from the `terra` package
 * Data can be returned as an object of class 'matrix' when using the argument `as.matrix = TRUE` in the S3 methods for objects of class 'default', 'SpatVector' and 'SpatRaster'
 
-# CHANGES IN BEHAVIOUR
+### CHANGES IN BEHAVIOUR
 
 * A new argument `server = ` is added to `get_chirps()` to indicate from which server the data will be requested. Options are 'CHC' the default and 'ClimateSERV'
 * Argument `operation = ` in `get_chirps()` is removed from the main Usage and is only required when `server = "ClimateSERV"`